updated license info
ATTENTION
This is the last commit to the svn repository!
The repository will be read-only after this operation!
We are migrating to Gitlab now.
Do NOT use the Gitlab repository until you get informed by either me or Martin!
We try to be quick but do not expect operation of the Gitlab repo before early next week!
ATTENTION