polishing and testing
element 11/12 (7/57 for MSC.Marc) have the same geometry type, i.e. both
are a hexahedral with 8 integration points/cells
Type 11 is linear (8 nodes), type 12 quadradic with reduced integration (20 nodes).
The modified definition ensures that the cell nodes 9-20 of element 11
are in the same order as the real nodes 9-20 of element 12. Real nodes
1-8 (corners) and 21-27 needed no modification.

Notes:
 * Documentation on https://damask.mpie.de/Documentation/ElementType is now outdated.
 * Element defition in MSC.Marc manual volume B (2001 version) is
   confusing because element numbering is sometimes clowise and
   sometimes counterclockwise. The latter one seems to be correct