delegate __getitem__ to numpy
might be not the fastest solution, but ensures that the behavior matches
that of np.ndarray.

Tests will be added once !1015 is merged because using random slices is
beneficial here. Once test coverage is good, we can try to implement
faster solutions involving mixed indexing (but Python 3.10 and 3.13 seem
to behave differently in corner cases involving "*()".