environment variable instead of preprocessor definition
makes it easier to control redirection to file.

Moved isaTTY to IO because it is only used there and needs access to a
(module) variable that encompasses whether files are redirected. Seems
also more logical to have it there because it is IO related.
Note: invoking "inquire" in the isaTTY function leads to a deadlock
(tested with Gfortran).