New `Table.unique` integrated into `from_table` methods of `GeomGrid` and `ConfigMaterial`.
In contrast to `numpy.unique`, the `Table.unique` function does not return the unique values sorted but in the order of their appearance in the input table. Apart from that, all other functionality is transferred, i.e., optional `return_index`, `return_inverse`, and `return_counts`.

NumPy 2.0.0 introduced a breaking change by returning multidimensional inverse values for `axis != None`. This has been reverted in \>2.0.0, but requires a fix in our code to be compatible with 2.0.0 (lack of that resulted in pipeline failure so far)!