Regrid for combined stretch and shear deformation
The origin, together with the three (x,y,z) edges of the originally hexahedral geometry, form a lattice of periodically equivalent points. Iff this lattice (dealing with a subset of it might be discussed later) has an orthogonal basis aligned with x,y,z, then regridding based on the hexahedral volume element spanned by that orthogonal basis is possible.