Correct use of SemVer for intermediate versions
Adding with a hyphen (-) to MAJOR.MINOR.PATCH is interpreted as a
pre-release, i.e. it is sorted before MAJOR.MINOR.PATCH.
The correct sign to append "build metadata" is a plus (+).
pip (Python) understands it.
Note that the string produced  by `git describe` is incompatible
and needs to be adjusted.
See also https://github.com/semver/semver/issues/200.