use IO_warning to inform about failure
this ensures that messages are seen in the case of MPI simulations
(stdout is disabled for rank != 0, stderr not).

extra guard needed in IO_warning/panel to avoid scrambled output. This
should not happen, but was reported at least for earlier Gfortran
versions:
https://fortran-lang.discourse.group/t/openmp-and-thread-safety-of-i-os-write-read/4567/19.
The resulting output (potentially incomplete ANSI escape codes)
are invalid UTF-8 which cannot be handled by Python's subprocess
library

closes #533