Support named mesh element groups for BC
Allow the use of named groups (labels) from the mesh file to
apply BC and/or material ID, instead of numbers (tag).

Using labels is still allowed for backwards compatibility.
If a group has a label, either of the tag or the label can be used to
reference it (but not both).

Add error checks for:

- Inexistent labels (i.e. trying to apply a BC onto a label not defined
  in the mesh file).
- No groups defined for material ID assignment.
  - Groups must be created; they may or may not be named.

Use the label for printing loadcase information, when appropriate.

Update/improve some loadcase/boundary condition error messages.